Add mapping external container id to internal functioanlity.
As we need to return the response to user before container is actually created we must have map between external id for user and internal id for docker. The approach will not scale horizontally(adding new instances) as the map is stored in memory. For such we could use Redis to store there the map.